Added in the 'getting started' docs
The 'getting started' docs are the ones that all users should be
familiar with before using any ZOOlander element - they are common to
all (or almost all) elements.
This is the initial content post of these.
(Once these are OK and you let me know about the process of adding
images I'll update these with images.